rq: add util
